+++ R/lp_getkBestRanked.R
@@ -1,8 +1,17 @@
-############################################################################
+###############################################################################################################################################
 #
-# get_k_best variant for ranked algorithm based on the LP
+# Murty's algorithm for k-best assignments
 #
-############################################################################
+# This version is executed when ranking is needed and when LP (Simplex) is used.
+#
+# @param matR Square matrix (N x N) in which values represent the weights.
+# @param k_bestR How many best scenarios should be returned. If by_rank = TRUE, this equals best ranks.
+# @param objectiveR Should the cost be minimized ('min') or maximized ('max')? Defaults to 'min'.
+# @param proxy_InfR What should be considered as a proxy for Inf? Defaults to 10e06; if objective = 'max' the sign is automatically reversed.
+#
+# @return A list with solutions and costs (objective values).
+#
+###############################################################################################################################################
 
 getkBestRankedLP <- function(matR, k_bestR = NULL, objectiveR = 'min', proxy_InfR = proxy_Inf) {
   
@@ -26,7 +35,7 @@ getkBestRankedLP <- function(matR, k_bestR = NULL, objectiveR = 'min', proxy_Inf
     
   }
   
-  if (k_bestR < 1) { stop("You have provided an invalid value for k_bestR.") }
+  if (k_bestR < 1) { stop("You have provided an invalid value for k_best.") }
   
   # Stripping the dimension names - column names need to be V1, V2, V3 .. in order to reconstruct the full matrix
   
@@ -186,15 +195,7 @@ getkBestRankedLP <- function(matR, k_bestR = NULL, objectiveR = 'min', proxy_Inf
     
     # Check fullObjs for the (remaining) optimal (minimum/maximum) cost, the next iteration uses it as starting basis
     
-    if (objectiveR == 'min') {
-      
-      idxOpt <- which.min(fullObjs)
-      
-    } else {
-      
-      idxOpt <- which.max(fullObjs)
-      
-    }
+    idxOpt <- if (objectiveR == 'min') which.min(fullObjs) else which.max(fullObjs)
     
     # Store the corresponding full matrix & related information into variables needed for each iteration
     
@@ -260,21 +261,7 @@ getkBestRankedLP <- function(matR, k_bestR = NULL, objectiveR = 'min', proxy_Inf
       warning(
         paste0(
           "There are ", n_possible, " possible solutions. Final solution has been found at rank number ",
-          length(all_solutions), " which is lower than the k_bestR specified; terminating here."
-        )
-      )
-      
-      break
-      
-    } else if (
-      
-      ( (length(all_solutions) == n_possible) | (length(unlist(all_objectives)) == n_possible) ) & (k_bestR > n_possible) 
-      
-    ) {
-      
-      warning(
-        paste0(
-          "There are only ", n_possible, " possible solutions; terminating earlier, stopping at rank ", length(all_solutions), "."
+          length(all_solutions), " which is lower than the k_best specified; terminating here."
         )
       )